Split doubled-cyber-dojo code into separate file [ci skip]
